Plot spatial cluster heatmap for each data batch

The function EachClusterSpaHeatMap is used to visualize clusters on spatial coordinates with the RGB colors for each data batch.

EachClusterSpaHeatMap(obj, ...)

# S3 method for list
EachClusterSpaHeatMap(
  obj,
  cols = NULL,
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(round(length(obj[[2]])/2), 2),
  no_axis = FALSE,
  pt_size = 2,
  point_alpha = 1,
  nrow.legend = NULL,
  base_size = 10,
  border_col = "gray",
  axis_names = c("Coord x", "Coord y"),
  no_guides = FALSE,
  no_axis_name = FALSE,
  legend.position = "bottom",
  common.legend = TRUE
)

# S3 method for SRTProject
EachClusterSpaHeatMap(
  obj,
  batch = NULL,
  cols = NULL,
  layout.dim = NULL,
  combine = TRUE,
  no_axis = TRUE,
  ...
)

Arguments

obj

A SRTProject object.

...

Arguments passed to other methods.

cols

a string vector, colors used in the plot.

title_name

a string, the title of plot.

combine

a logical value, whether plot all on a figure. If TRUE, all figures are plotted; otherwise, return a list with each plot as component. TRUE by default.

layout.dim

a 2 vector, the dimension in the layout of plots when combine = TRUE.

no_axis

a logical value, whether displys the axis.

batch

a string, number or NULL, specify the batch name to plot, default as NULL implying all batches plotted.