Plot the embeddings for each SRT data

Plot the embeddings for each SRT data

EachEmbedPlot(obj, reduction = NULL, batch = NULL, ...)

# S3 method for SpatialExperiment
EachEmbedPlot(
  obj,
  reduction = NULL,
  colour_by = NULL,
  batch = NULL,
  combine = TRUE,
  base_size = 15,
  layout.dim = NULL,
  common.legend = TRUE,
  legend.position = "right",
  add_border_box = FALSE,
  no_axis = TRUE,
  ...
)

Arguments

obj

an object with class SpatialExperiment .

reduction

a string, speficify which reduction slot is used to plot. If reduction=NULL, the last one is used.

batch

a string vector, denoting the names of batches to be plotted, or a integer vector denoting the ID numbers of batches to be plotted.

...

other arguments that pass to other function.

colour_by

a string, specify the colname in cell metadata to be mapped to colors.

combine

a logical value, whether put all plots in one figure.

base_size

a positive real, the baseline text size.

layout.dim

a 2-dimension integer vector, specify the layout of figure.

common.legend

a logical value, whether use a common legend for all plots.

legend.position

a string, the position of legend.

add_border_box

a logical value, whether add the border box.

no_axis

a logical value, whether display the axis.

Value

return a ggplot object of a list of ggplot objects.