Joint embedding, clustering and alignment for SRT data by fitting a PRECAST model

Joint embedding, clustering and alignment for SRT data by fitting a PRECAST model

Integrate_PRECAST(SRTProj, K = NULL, q = 15, ...)

Arguments

SRTProj

an object named "SRTProject". The object SRTProject is created by CreateSRTProject.

K

An optional integer or integer vector, specify the candidates of number of clusters. if K=NULL, it will be set to 4~12.

q

An optional integer, specify the number of low-dimensional embeddings to extract in PRECAST.

...

other parameters pass to model_set_PRECAST.

References

Wei Liu, Liao, X., Luo, Z. et al, Jin Liu* (2023). Probabilistic embedding, clustering, and alignment for integrating spatial transcriptomics data with PRECAST. Nature Communications, 14, 296