Joint clustering and alignment for SRT data by fitting an iSC.MEB model

Joint clustering and alignment for SRT data by fitting an iSC.MEB model

Integrate_iSCMEB(SRTProj, K = NULL, reduction = "PCA", ...)

Arguments

SRTProj

an object named "SRTProject". The object SRTProject is created by CreateSRTProject.

K

an optional integer or integer vector, specify the candidates of number of clusters. if K=NULL, it will be set to 4~12.

reduction

a string, specify the reduction slot name in SRTProj for fitting iSC.MEB model.

...

other parameters pass to model_set_iSCMEB.

References

iSC.MEB: an R package for multi-sample spatial clustering analysis of spatial transcriptomics data, Bioinformatics Advances, 2023, vbad019