Barplot for the enrichment terms

Barplot for the enrichment terms

barPlot_enrich(
  top_dat,
  source = "Ont",
  term_name = "Term",
  nlog10P = "nlog10P",
  bar_width = 0.8,
  base_size = 20,
  font_family = "serif",
  cols = (ggthemes::canva_pal())(4)
)

Arguments

top_dat

a data frame with the information of enrichment terms

source

a string, the source of database.

term_name

a string, the name of y-axis.

nlog10P

a string, the colname in top_dat that represents the negative log10 p-values.

bar_width

a positive real, the width of bars.

base_size

a positive real, the baseline text size in plot.

font_family

a string, the font.

cols

a string vector, the used colors in plot.

Value

return a ggplot object.