Set parameters for SC-MEB model

Prepare parameters setup for SC-MEB model fitting.

model_set_SCMEB(
  beta_grid = seq(0.2, 4, by = 0.2),
  coreNum = 1,
  maxIter_ICM = 6,
  maxIter = 30,
  verbose = TRUE,
  seed = 1
)

Arguments

beta_grid

a numeric vector, specify the smoothness parameter of Random Markov Field. The default is seq(0,4,0.2).

coreNum

an integer value to decide how many cores are used to run SC-MEB model in parallel.

maxIter_ICM

the maximum iteration of ICM step, default as 6.

maxIter

the maximum iteration of ICM-EM algorithm. The default is 30.

verbose

a logical value, whether output the information in iteration.

seed

a postive integer, the random seed to be set.

Examples

model_set_SCMEB(coreNum=5)
#> $coreNum
#> [1] 5
#> 
#> $beta_grid
#>  [1] 0.2 0.4 0.6 0.8 1.0 1.2 1.4 1.6 1.8 2.0 2.2 2.4 2.6 2.8 3.0 3.2 3.4 3.6 3.8
#> [20] 4.0
#> 
#> $maxIter_ICM
#> [1] 6
#> 
#> $maxIter
#> [1] 30
#> 
#> $verbose
#> [1] TRUE
#> 
#> $seed
#> [1] 1
#>