SpatialExperiment object to Seurat object

Transfer SpatialExperiment object to a Seurat object for preparation for DR.SC model fitting.

spe2seurat(spe, verbose= TRUE)

Arguments

spe

a SpatialExperiment object, at least including the raw gene count expression matrix ans sptial coordinates.

verbose

an optional logical value, whether output the information.

Value

Return a Seurat object, where the spatial coordinates information is saved in the metadata of Seurat, named "row" and "col".

Author

Wei Liu

Note

nothing

See also

Examples

  dir <- system.file(
  file.path("extdata", "10xVisium", "section1"),
  package = "SpatialExperiment")

  # read in counts
  fnm <- file.path(dir, "raw_feature_bc_matrix")
  sce <- DropletUtils::read10xCounts(fnm)

  # read in image data
  img <- readImgData(
    path = file.path(dir, "spatial"),
    sample_id="foo")

  # read in spatial coordinates
  fnm <- file.path(dir, "spatial", "tissue_positions_list.csv")
  xyz <- read.csv(fnm, header = FALSE,
                  col.names = c(
                    "barcode", "in_tissue", "array_row", "array_col",
                    "pxl_row_in_fullres", "pxl_col_in_fullres"))

  # construct observation & feature metadata
  rd <- S4Vectors::DataFrame(
    symbol = rowData(sce)$Symbol)

  # construct 'SpatialExperiment'
  (spe <- SpatialExperiment(
    assays = list(counts = assay(sce)),
    colData = colData(sce), rowData = rd, imgData = img,
    spatialData=DataFrame(xyz),
    spatialCoordsNames=c("pxl_col_in_fullres", "pxl_row_in_fullres"),
    sample_id="foo"))
#> Found more than one class "SpatialImage" in cache; using the first, from namespace 'SeuratObject'
#> Also defined by 'SpatialExperiment'
#> class: SpatialExperiment 
#> dim: 50 50 
#> metadata(0):
#> assays(1): counts
#> rownames(50): ENSMUSG00000051951 ENSMUSG00000089699 ...
#>   ENSMUSG00000005886 ENSMUSG00000101476
#> rowData names(1): symbol
#> colnames: NULL
#> colData names(3): Sample Barcode sample_id
#> Found more than one class "SpatialImage" in cache; using the first, from namespace 'SeuratObject'
#> Also defined by 'SpatialExperiment'
#> reducedDimNames(0):
#> mainExpName: NULL
#> Found more than one class "SpatialImage" in cache; using the first, from namespace 'SeuratObject'
#> Also defined by 'SpatialExperiment'
#> altExpNames(0):
#> spatialData names(4) : barcode in_tissue array_row array_col
#> spatialCoords names(2) : pxl_col_in_fullres pxl_row_in_fullres
#> imgData names(4): sample_id image_id data scaleFactor

  colnames(spe) <- paste0("spot", 1:ncol(spe))

  seu <- spe2seurat(spe)
#> Error in spe2seurat(spe): could not find function "spe2seurat"
  head(seu)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'seu' not found

  ## Fit DR-SC model
  library(DR.SC)
  seu <- NormalizeData(seu)
#> Error in NormalizeData(seu): could not find function "NormalizeData"
  seu <- FindVariableFeatures(seu)
#> Error in FindVariableFeatures(seu): could not find function "FindVariableFeatures"
  seu <- DR.SC(seu = seu, K=4)
#> Error in DR.SC(seu = seu, K = 4): object 'seu' not found